Added support for non-integer radius to //superpickaxe recursive and improved performance a bit by making the recurse method static.

+++ src/main/java/com/sk89q/worldedit/commands/SuperPickaxeCommands.java
@@ -88,7 +88,7 @@ public class SuperPickaxeCommands {
             throws WorldEditException {
 
         LocalConfiguration config = we.getConfiguration();
-        int range = args.getInteger(0);
+        double range = args.getDouble(0);
         
         if (range > config.maxSuperPickaxeSize) {
             player.printError("Maximum range: " + config.maxSuperPickaxeSize);
